requirements.txt
# Kevin AI - SOP Automation Platform
# Python Requirements
# Version: 2.0
# Date: January 6, 2026

# =============================================================================
# CORE FRAMEWORK
# =============================================================================
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings==2.12.0

# =============================================================================
# UI FRAMEWORK
# =============================================================================
streamlit==1.40.2
# gradio==5.9.1  # Optional - use streamlit instead

# =============================================================================
# LANGCHAIN & AI
# =============================================================================
langchain==1.2.0
langchain-openai==1.1.6
langchain-community==0.4.1
langchain-anthropic==0.3.5
langchain-core==1.2.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
openai==2.14.0
anthropic==0.75.0

# =============================================================================
# VECTOR STORE & EMBEDDINGS
# =============================================================================
faiss-cpu==1.13.2
tiktoken==0.12.0

# =============================================================================
# DOCUMENT PROCESSING
# =============================================================================
pymupdf==1.26.7
python-docx==1.2.0
pillow==11.3.0
pytesseract==0.3.13
pdf2image==1.17.0

# =============================================================================
# DATA PROCESSING
# =============================================================================
pandas==2.2.3
numpy==2.3.5
openpyxl==3.1.5

# =============================================================================
# HTTP & NETWORKING
# =============================================================================
python-dotenv==1.2.1
requests==2.32.5
aiohttp==3.13.3
httpx==0.28.1
websockets==14.2

# =============================================================================
# MONITORING & LOGGING
# =============================================================================
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# =============================================================================
# TESTING
# =============================================================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
black==24.12.0
mypy==1.13.0
ruff==0.14.10

# =============================================================================
# ADDITIONAL DEPENDENCIES
# =============================================================================
# For JSON handling
orjson==3.11.5

# For async operations
anyio==4.12.0
aiofiles==23.2.1

# For CLI
typer==0.21.0
rich==14.2.0

# For file handling
fsspec==2025.12.0

# For API documentation
markdown-it-py==4.0.0

# Document Generation
reportlab==4.2.5
python-pptx==1.0.2

# Mermaid to Image Conversion
playwright==1.49.1
